Refactoring and profile page MVP
- Refactored components
- Made the profile page posts connect to firestore
