+++ wall4comment/src/API/auth.js
@@ -6,6 +6,7 @@ const auth = getAuth(app);
 let AuthContext;
 function AuthWrapper({ children }) {
     AuthContext = createContext(null);
+    // user is set to false so that if there is no user it will be set to null
     const [user, setUser] = useState(false);
 
     useEffect(() => {
